power: qpnp-fg: Return error properly when IMA read/write fails
In the SRAM masked write we read the register first, make
modification to the value and write it back. If the read fails
we bail out of the masked write function.

However there is a bug in the read api where the error code is
not properly returned. This causes the masked write to proceed
with a failed read and corruption ensues.

A similar bug is present in write api too where the error is not
returned correctly to the caller. Fix this. While at it, take the
opportunity to improve debug logs.

Bug: 33775329
Change-Id: I4ad58b8e51212ad43a12cdc611c621d1dd62df20
Signed-off-by: Subbaraman Narayanamurthy <subbaram@quicinc.com>
Signed-off-by: ahmedradaideh <ahmed.radaideh@gmail.com>